+++ src/aui-form-validator/js/aui-form-validator.js
@@ -958,10 +958,8 @@ var FormValidator = A.Component.create({
          */
         _defErrorFieldFn: function(event) {
             var instance = this,
-                ancestor,
                 field,
                 label,
-                nextSibling,
                 stackContainer,
                 target,
                 validator;
@@ -977,21 +975,8 @@ var FormValidator = A.Component.create({
 
                 stackContainer = instance.getFieldStackErrorContainer(field);
 
-                nextSibling = field.get('nextSibling');
-
-                if (nextSibling && nextSibling.get('nodeType') === 3) {
-                    ancestor = field.ancestor();
-
-                    if (ancestor) {
-                        if (ancestor.hasClass(label)) {
-                            target = nextSibling;
-                        }
-                        else if (A.FormValidator.isCheckable(target)) {
-                            var label = ancestor.previous('.' + label);
-
-                            target = label;
-                        }
-                    }
+                if (A.FormValidator.isCheckable(target)) {
+                    target = field.ancestor('.' + CSS_ERROR).get('lastChild');
                 }
 
                 target.placeAfter(stackContainer);
@@ -1067,7 +1052,26 @@ var FormValidator = A.Component.create({
         },
 
         /**
-         * TODO. Wanna help? Please send a Pull Request.
+         * Finds the label text of a field if existing.
+         *
+         * @method _findFieldLabel
+         * @param field
+         * @return {String}
+         */
+        _findFieldLabel: function(field) {
+            var labelCssClass = '.' + this.get('labelCssClass'),
+                label =  A.one('label[for=' + field.get('id') + ']') ||
+                    field.ancestor().previous(labelCssClass) ||
+                    field.ancestor('.' + CSS_ERROR).one(labelCssClass);
+
+            if (label) {
+                return label.get('text');
+            }
+        },
+
+        /**
+         * Sets the error/success CSS classes based on the validation of a
+         * field.
          *
          * @method _highlightHelper
          * @param field