+++ src/aui-form-validator/HISTORY.md
@@ -4,4 +4,5 @@ AUI Form Validator
 @VERSION@
 ------
 
+	* #AUI-1585 aui-form-validator error breaks radio button layout
 	* #AUI-965 Fixed issue with select box items and displaying error messages